requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
pandas==2.1.4
numpy==1.26.2
Pillow==10.1.0
Flask==3.0.0 